requirements.txt
Django==4.2.7
djangorestframework==3.14.0
PyMySQL==1.1.0
cryptography==41.0.5
celery==5.3.6
pydantic-settings==2.1.0
redis==5.0.1
async_timeout==4.0.3
